requirements.txt
streamlit
tensorflow
numpy
pillow
requests
gdown
